requirements.txt
requests>=2.26.0
pytest>=7.0.0
pytest-allure-adaptor>=1.0.7
PyYAML>=6.0
allure-pytest>=2.9.45
jsonpath>=0.82
colorlog>=6.4.1

